working on normal again, almost there

+++ includes/sphere.h
@@ -11,7 +11,7 @@ typedef struct s_sphere
 }	t_sphere;
 
 
-int			hit_sphere(const t_sphere *sphere, const t_ray *ray);
+double hit_sphere(const t_sphere *sphere, const t_ray *ray);
 t_sphere	*new_sphere(t_vec3 *pos, double diameter, t_vec3 *color);
 void		free_sphere(t_sphere *sphere);
